`bookmarksToSuggestions` tags each suggestion with the bookmark
`bookmarksToSuggestions` generates a suggestion for every bookmark.
But since it ignores bookmark folders and bookmarklets, `s.length` may
be shorter than `b.length`.

`searchInput` assumed that there was a 1:1 match between every
bookmark and suggestion. Now, each suggestion is tagged with a
`bookmark` property that points to the bookmark it was generated from.

Unfortunately, `chrome.omnibox.sendSuggestions` rejects unexpected
properties, so it needs to `clean` the `bookmark` property from the
suggestions before `suggest` is called.